Add Memory-Map interface.
Hardware devices pass the cpu a base addres, length, and callback
function. If a memory access occurs between base and base+len,
then the callback function is called. The signature of the
callback function is:

callback: (addr, val) ->

Addr is the offset from the base address specified. If val is
defined, then this is a write operation, otherwise it is a read.