Add a simple debug system.
In source files, the assembler will detect special comment lines of the
form:
;! javascript_code_here;

The javascript statements will be packaged up and the emulator will
eval them at runtime. The context in which these statements will be
evaled will contain a variable called "cpu" containing the cpu
instance. Also defined is a function called "assert" which takes one
boolean parameter. If this parameter is false, then (on the command
line), the emulator will halt and display the error message.