Fix instruction decode issue.
There were missing undefined entries in the OpCode table which caused
valid instructions to be decoded as invalid.

Also clamp all register writes to be only 16 bits so that it is
impossible to write a larger value to them.

Add some minor fixes to the sti instruction, and removed some checks in
add that are no longer needed with clamped register writes.